Improve vis tests descriptions
