+++ cartoframes/vis/__init__.py
@@ -5,10 +5,12 @@ from .map import Map
 from .layer import Layer
 from .style import Style
 from .source import Source
+from . import defaults
 
 
 __all__ = [
     'basemaps',
+    'defaults',
     'Map',
     'Layer',
     'Source',
