+++ cartoframes/vis/defaults.py
