+++ test/test_dataset.py
@@ -10,7 +10,7 @@ import warnings
 from carto.exceptions import CartoException
 
 from cartoframes.context import CartoContext
-from cartoframes.datasets import Dataset
+from cartoframes import Dataset
 from cartoframes.columns import normalize_name
 
 from utils import _UserUrlLoader
