+++ test/vis/test_layer.py
@@ -2,36 +2,37 @@ import unittest
 from cartoframes import vis
 
 
-class TestLayers(unittest.TestCase):
+class TestLayer(unittest.TestCase):
     def test_is_layer_defined(self):
+        """vis.Layer"""
         self.assertNotEqual(vis.Layer, None)
 
     def test_initialization_objects(self):
-        """should initialize layer attributes"""
-        layer = vis.Layer(vis.Source('layer_source'), vis.Style())
+        """vis.Layer should initialize layer attributes"""
+        layer = vis.Layer(vis.Source('layer_source'))
 
         self.assertEqual(layer.orig_query, 'SELECT * FROM layer_source')
         self.assertFalse(layer.is_basemap)
-        self.assertEqual(layer.style, '')
+        self.assertTrue(isinstance(layer.style, vis.Style))
         self.assertEqual(layer.viz, '')
         self.assertIsNone(layer.interactivity)
         self.assertIsNone(layer.legend)
 
     def test_initialization_simple(self):
-        """should initialize layer attributes"""
+        """vis.Layer should initialize layer attributes"""
         layer = vis.Layer('layer_source', '')
 
         self.assertEqual(layer.orig_query, 'SELECT * FROM layer_source')
         self.assertFalse(layer.is_basemap)
-        self.assertEqual(layer.style, '')
+        self.assertTrue(isinstance(layer.style, vis.Style))
         self.assertEqual(layer.viz, '')
         self.assertIsNone(layer.interactivity)
         self.assertIsNone(layer.legend)
 
 
-class TestLayersStyle(unittest.TestCase):
+class TestLayerStyle(unittest.TestCase):
     def test_style_dict(self):
-        """should set the style when it is a dict"""
+        """vis.Layer style should set the style when it is a dict"""
         layer = vis.Layer(
             'layer_source',
             {
@@ -53,7 +54,7 @@ class TestLayersStyle(unittest.TestCase):
         self.assertTrue('strokeWidth: 1' in layer.viz)
 
     def test_style_str(self):
-        """should set the style when it is a dict"""
+        """vis.Layer style should set the style when it is a dict"""
         layer = vis.Layer(
             'layer_source',
             """
@@ -72,12 +73,12 @@ class TestLayersStyle(unittest.TestCase):
         self.assertTrue('strokeColor: black' in layer.viz)
         self.assertTrue('strokeWidth: 1' in layer.viz)
 
-    @raises(ValueError)
     def test_style_dict_valid_properties(self):
-        """should set only the valid properties"""
-        layer = vis.Layer(
-            'layer_source',
-            {
-                'invalid': 1
-            }
-        )
+        """vis.Layer style should set only the valid properties"""
+        with self.assertRaises(ValueError):
+            vis.Layer(
+                'layer_source',
+                {
+                    'invalid': 1
+                }
+            )
